backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
pydantic-settings
python-jose[cryptography]
python-docx
markdown2
psycopg2-binary
httpx
